+++ GRDB Tests/Public/Record/MinimalPrimaryKeyRowIDTests.swift
@@ -256,8 +256,8 @@ class MinimalPrimaryKeyRowIDTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = MinimalRowID()
                 record.id = 123456
-                let changes = try record.delete(db)
-                XCTAssertEqual(changes.changedRowCount, 0)
+                let deleted = try record.delete(db)
+                XCTAssertFalse(deleted)
             }
         }
     }
@@ -267,8 +267,8 @@ class MinimalPrimaryKeyRowIDTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = MinimalRowID()
                 try record.insert(db)
-                let changes = try record.delete(db)
-                XCTAssertEqual(changes.changedRowCount, 1)
+                let deleted = try record.delete(db)
+                XCTAssertTrue(deleted)
                 
                 let row = Row.fetchOne(db, "SELECT * FROM minimalRowIDs WHERE id = ?", arguments: [record.id])
                 XCTAssertTrue(row == nil)
@@ -281,10 +281,10 @@ class MinimalPrimaryKeyRowIDTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = MinimalRowID()
                 try record.insert(db)
-                var changes = try record.delete(db)
-                XCTAssertEqual(changes.changedRowCount, 1)
-                changes = try record.delete(db)
-                XCTAssertEqual(changes.changedRowCount, 0)
+                var deleted = try record.delete(db)
+                XCTAssertTrue(deleted)
+                deleted = try record.delete(db)
+                XCTAssertFalse(deleted)
             }
         }
     }