+++ GRDB Tests/Public/Record/MinimalPrimaryKeySingleTests.swift
@@ -252,8 +252,8 @@ class MinimalPrimaryKeySingleTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = MinimalSingle()
                 record.UUID = "theUUID"
-                let changes = try record.delete(db)
-                XCTAssertEqual(changes.changedRowCount, 0)
+                let deleted = try record.delete(db)
+                XCTAssertFalse(deleted)
             }
         }
     }
@@ -264,8 +264,8 @@ class MinimalPrimaryKeySingleTests: GRDBTestCase {
                 let record = MinimalSingle()
                 record.UUID = "theUUID"
                 try record.insert(db)
-                let changes = try record.delete(db)
-                XCTAssertEqual(changes.changedRowCount, 1)
+                let deleted = try record.delete(db)
+                XCTAssertTrue(deleted)
                 
                 let row = Row.fetchOne(db, "SELECT * FROM minimalSingles WHERE UUID = ?", arguments: [record.UUID])
                 XCTAssertTrue(row == nil)
@@ -279,10 +279,10 @@ class MinimalPrimaryKeySingleTests: GRDBTestCase {
                 let record = MinimalSingle()
                 record.UUID = "theUUID"
                 try record.insert(db)
-                var changes = try record.delete(db)
-                XCTAssertEqual(changes.changedRowCount, 1)
-                changes = try record.delete(db)
-                XCTAssertEqual(changes.changedRowCount, 0)
+                var deleted = try record.delete(db)
+                XCTAssertTrue(deleted)
+                deleted = try record.delete(db)
+                XCTAssertFalse(deleted)
             }
         }
     }